// src/api-client/models/verification-code-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spring-Adam-Store
 * Api Service
 *
 * The version of the OpenAPI document: v1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface VerificationCodeResponse
 */
export interface VerificationCodeResponse {
    /**
     * 
     * @type {string}
     * @memberof VerificationCodeResponse
     */
    'email'?: string;
    /**
     * 
     * @type {string}
     * @memberof VerificationCodeResponse
     */
    'verificationCode'?: string;
    /**
     * 
     * @type {number}
     * @memberof VerificationCodeResponse
     */
    'ttl'?: number;
}