// src/api-client/models/redis-forgot-password-token.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spring-Adam-Store
 * Api Service
 *
 * The version of the OpenAPI document: v1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface RedisForgotPasswordToken
 */
export interface RedisForgotPasswordToken {
    /**
     * 
     * @type {string}
     * @memberof RedisForgotPasswordToken
     */
    'forgotPasswordToken'?: string;
    /**
     * 
     * @type {string}
     * @memberof RedisForgotPasswordToken
     */
    'email'?: string;
    /**
     * 
     * @type {number}
     * @memberof RedisForgotPasswordToken
     */
    'ttl'?: number;
}